+++ packages/ethereum-contracts/contracts/interfaces/agreements/IConstantFlowAgreementV1.sol
@@ -144,6 +144,42 @@ abstract contract IConstantFlowAgreementV1 is ISuperAgreement {
     ) external virtual returns(bytes memory newCtx);
 
     /**
+     * @dev msgSender from `ctx` increases flow rate allowance for the `flowOperator` by `addedFlowRateAllowance`
+     * @dev if `addedFlowRateAllowance` is negative, we revert with CFA_ACL_NO_NEGATIVE_ALLOWANCE
+     * @param token Super token address
+     * @param flowOperator The permission grantee address
+     * @param permissions A bitmask representation of the granted permissions
+     * @param addedFlowRateAllowance The flow rate allowance delta
+     * @param ctx Context bytes (see ISuperfluid.sol for Context struct)
+     * @return newCtx The new context bytes
+     */
+    function increaseFlowAllowanceWithPermissions(
+        ISuperfluidToken token,
+        address flowOperator,
+        uint8 permissions,
+        int96 addedFlowRateAllowance,
+        bytes calldata ctx
+    ) external virtual returns(bytes memory newCtx);
+
+    /**
+     * @dev msgSender from `ctx` decreases flow rate allowance for the `flowOperator` by `subtractedFlowRateAllowance`
+     * @dev if `subtractedFlowRateAllowance` is negative, we revert with CFA_ACL_NO_NEGATIVE_ALLOWANCE
+     * @param token Super token address
+     * @param flowOperator The permission grantee address
+     * @param permissions A bitmask representation of the granted permissions
+     * @param subtractedFlowRateAllowance The flow rate allowance delta
+     * @param ctx Context bytes (see ISuperfluid.sol for Context struct)
+     * @return newCtx The new context bytes
+     */
+    function decreaseFlowAllowanceWithPermissions(
+        ISuperfluidToken token,
+        address flowOperator,
+        uint8 permissions,
+        int96 subtractedFlowRateAllowance,
+        bytes calldata ctx
+    ) external virtual returns(bytes memory newCtx);
+
+    /**
      * @dev msgSender from `ctx` grants `flowOperator` all permissions with flowRateAllowance as type(int96).max
      * @param token Super token address
      * @param flowOperator The permission grantee address